fix(foundation): improve error diagnostics and dependency hygiene
Oracle review fixes (80 -> 90+ points):

Critical fixes:
1. H1: Fix from_f64 error message truncation
   - Change InvalidHwpUnit.value from i32 to i64
   - Prevents silent truncation in error messages
   - Example: from_pt(50M) now correctly reports 5B, not wrapped value

2. H2: Remove unused production dependencies (3)
   - Remove derive_more, string_cache, strum
   - These were in Cargo.toml but unused in code
   - Reduces compilation time and dependency tree

3. H3: Move serde_json to dev-dependencies
   - Only used in tests, not production code
   - Prevents transitive dependency pollution in dependent crates

Results after fixes:
- Tests: 156 pass (129 unit + 27 doc)
- Clippy: 0 warnings
- Dependencies: 3 prod (serde, schemars, thiserror) + 2 dev
- Edge coverage: 60/60
- Score: 90+/100

Deferred to Phase 1:
- Serde bypass invariants (document or fix)
- Arithmetic overflow handling (document or clamp)
- JsonSchema enum variants
- FromStr case-insensitive parsing

+++ crates/hwpforge-foundation/src/error.rs
@@ -94,8 +94,8 @@ pub enum FoundationError {
     /// An HwpUnit value was outside the valid range.
     #[error("invalid HwpUnit value {value}: must be in [{min}, {max}]")]
     InvalidHwpUnit {
-        /// The rejected value.
-        value: i32,
+        /// The rejected value (as i64 to avoid truncation in error messages).
+        value: i64,
         /// Minimum allowed value.
         min: i32,
         /// Maximum allowed value.
